feat: Add auth endpoints

+++ database/migrations/2024_08_21_090910_create_student_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('firstName');
             $table->string('lastName');
             $table->string('location');
-            $table->integer('phoneNumber');
+            $table->bigInteger('phoneNumber');
             $table->timestamps();
         });
     }
